Clean number/integer/strings encoder/decoder options
