+++ src/options/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(sourcemeta_jsonbinpack_options INTERFACE
   include/jsonbinpack/options/enum.h
-  include/jsonbinpack/options/number.h
+  include/jsonbinpack/options/integer.h
   include/jsonbinpack/options/string.h)
 target_link_libraries(sourcemeta_jsonbinpack_options INTERFACE
   sourcemeta_jsontoolkit_json)
